Remove time package
