+++ ring/ticker_test.go
@@ -1,0 +1,34 @@
+package ring
+
+import (
+	"testing"
+	"time"
+)
+
+func TestNewDisableableTicker_Enabled(t *testing.T) {
+	stop, ch := newDisableableTicker(10 * time.Millisecond)
+	defer stop()
+
+	time.Sleep(100 * time.Millisecond)
+
+	select {
+	case <-ch:
+		break
+	default:
+		t.Error("ticker should have ticked when enabled")
+	}
+}
+
+func TestNewDisableableTicker_Disabled(t *testing.T) {
+	stop, ch := newDisableableTicker(0)
+	defer stop()
+
+	time.Sleep(100 * time.Millisecond)
+
+	select {
+	case <-ch:
+		t.Error("ticker should not have ticked when disabled")
+	default:
+		break
+	}
+}
