--- time/time.go
@@ -1,96 +1,0 @@
-package time
-
-import (
-	"math"
-	"math/rand"
-	"net/http"
-	"strconv"
-	"time"
-
-	"github.com/prometheus/common/model"
-	"github.com/weaveworks/common/httpgrpc"
-)
-
-const (
-	nanosecondsInMillisecond = int64(time.Millisecond / time.Nanosecond)
-)
-
-func ToMillis(t time.Time) int64 {
-	return t.UnixNano() / nanosecondsInMillisecond
-}
-
-// FromMillis is a helper to turn milliseconds -> time.Time
-func FromMillis(ms int64) time.Time {
-	return time.Unix(0, ms*nanosecondsInMillisecond)
-}
-
-// FormatTimeMillis returns a human readable version of the input time (in milliseconds).
-func FormatTimeMillis(ms int64) string {
-	return FromMillis(ms).String()
-}
-
-// FormatTimeModel returns a human readable version of the input time.
-func FormatTimeModel(t model.Time) string {
-	return FromMillis(int64(t)).String()
-}
-
-// ParseTime parses the string into an int64, milliseconds since epoch.
-func ParseTime(s string) (int64, error) {
-	if t, err := strconv.ParseFloat(s, 64); err == nil {
-		s, ns := math.Modf(t)
-		ns = math.Round(ns*1000) / 1000
-		tm := time.Unix(int64(s), int64(ns*float64(time.Second)))
-		return ToMillis(tm), nil
-	}
-	if t, err := time.Parse(time.RFC3339Nano, s); err == nil {
-		return ToMillis(t), nil
-	}
-	return 0, httpgrpc.Errorf(http.StatusBadRequest, "cannot parse %q to a valid timestamp", s)
-}
-
-// DurationWithJitter returns random duration from "input - input*variance" to "input + input*variance" interval.
-func DurationWithJitter(input time.Duration, variancePerc float64) time.Duration {
-	// No duration? No jitter.
-	if input == 0 {
-		return 0
-	}
-
-	variance := int64(float64(input) * variancePerc)
-	if variance == 0 {
-		// Values too low
-		return input
-	}
-
-	jitter := rand.Int63n(variance*2) - variance
-
-	return input + time.Duration(jitter)
-}
-
-// DurationWithPositiveJitter returns random duration from "input" to "input + input*variance" interval.
-func DurationWithPositiveJitter(input time.Duration, variancePerc float64) time.Duration {
-	// No duration? No jitter.
-	if input == 0 {
-		return 0
-	}
-
-	variance := int64(float64(input) * variancePerc)
-	if variance == 0 {
-		// Values too low
-		return input
-	}
-
-	jitter := rand.Int63n(variance)
-
-	return input + time.Duration(jitter)
-}
-
-// NewDisableableTicker essentially wraps NewTicker but allows the ticker to be disabled by passing
-// zero duration as the interval. Returns a function for stopping the ticker, and the ticker channel.
-func NewDisableableTicker(interval time.Duration) (func(), <-chan time.Time) {
-	if interval == 0 {
-		return func() {}, nil
-	}
-
-	tick := time.NewTicker(interval)
-	return func() { tick.Stop() }, tick.C
-}
